Create import_nps.r
Fuction to import NPS data to the global environment in R.
input: directory
output: nps_1#_$ where # is the last digit of the year and $ is the first digit of the table name.

Example. 
import_nps("~/NPS-2014")
import_nps("~/NPS-2015")
import_nps("~/NPS-2016")

-> output are nps_14_2, nps_14_3, nps_14_4, nps_14_5, nps_15_2, nps_15_3, nps_15_4, nps_15_5, nps_16_2, nps_16_3, nps_16_4, nps_16_5.
To be specific, nps_14_2 means 200table of year 2014.